+++ Runtime/Settings/OnInitErrorHandler.cs
@@ -1,4 +1,4 @@
 ﻿namespace AffiseAttributionLib.Settings
 {
-    public delegate void OnInitErrorHandler(string e);
+    public delegate void OnInitErrorHandler(string error);
 }